رفع خطاهای کامپایل Dart — تغییرات شکست‌دهنده Flutter 3.44 + مشکلات Drift
Flutter 3.44 API Changes:
- CardTheme → CardThemeData
- DialogTheme → DialogThemeData
- TabBarTheme → TabBarThemeData

اصلاح مشکلات کد:
- _CustomerFormDialog → CustomerFormDialog (public, قابل import)
- افزودن initialValue به CurrencyInput
- isSmallerThanExp → isSmallerThan (drift 2.27.0 API)
- اضافه کردن cast صریح (as double) برای FutureOr<double> در DAO‌ها

اصلاح وابستگی:
- drift ^2.18.0 → ^2.28.0 (برای سازگاری با drift_flutter 0.2.7)

+++ lib/data/local/daos/invoices_dao.dart
@@ -124,6 +124,6 @@ class InvoicesDao extends DatabaseAccessor<AppDatabase>
   /// جمع فروش امروز
   Future<double> getTodaySalesTotal() async {
     final invoices = await getTodayInvoices();
-    return invoices.fold(0.0, (sum, inv) => sum + inv.finalAmount);
+    return invoices.fold(0.0, (sum, inv) => sum + (inv.finalAmount as double));
   }
 }
